+++ src/DotNetProjectFile.Analyzers/Text/Globbing/Group.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.IO.Globbing;
+namespace DotNetProjectFile.Text.Globbing;
 
 internal sealed class Group(IReadOnlyList<Segement> segments) : Segement
 {
