+++ src/DotNetProjectFile.Analyzers/Text/Globbing/Literal.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.IO.Globbing;
+namespace DotNetProjectFile.Text.Globbing;
 
 internal sealed class Literal(string text) : Segement
 {
