+++ src/DotNetProjectFile.Analyzers/Text/Globbing/NotSequense.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.IO.Globbing;
+namespace DotNetProjectFile.Text.Globbing;
 
 internal sealed class NotSequence(string options) : Segement
 {
