+++ src/DotNetProjectFile.Analyzers/Text/Globbing/RecursiveWildcard.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.IO.Globbing;
+namespace DotNetProjectFile.Text.Globbing;
 
 internal sealed class RecursiveWildcard : Segement
 {
