+++ src/DotNetProjectFile.Analyzers/Text/Globbing/Sequence.cs
@@ -1,4 +1,4 @@
-namespace DotNetProjectFile.IO.Globbing;
+namespace DotNetProjectFile.Text.Globbing;
 
 internal sealed class Sequence(string options) : Segement
 {
